+++ Viewer.hpp
@@ -6,7 +6,7 @@ namespace haste {
 class Viewer : public Technique {
 public:
     Viewer(
-        const vector<vec4>& data,
+        const vector<dvec4>& data,
         size_t width,
         size_t height);
 
@@ -18,7 +18,7 @@ public:
     string name() const override;
 
 private:
-    vector<vec4> _data;
+    vector<dvec4> _data;
     size_t _width;
     size_t _height;
 };
