+++ framework.hpp
@@ -32,7 +32,7 @@ class Framework {
 public:
     virtual ~Framework();
 
-    virtual void render(size_t width, size_t height, glm::vec4* data) = 0;
+    virtual void render(size_t width, size_t height, glm::dvec4* data) = 0;
 
     virtual void updateUI(
         size_t width,
@@ -40,7 +40,7 @@ public:
         const glm::vec4* data,
         double elapsed) = 0;
 
-    virtual void postproc(glm::vec4* dst, const glm::vec4* src, size_t width, size_t height);
+    virtual void postproc(glm::vec4* dst, const glm::dvec4* src, size_t width, size_t height);
     virtual bool updateScene();
 
     int run(size_t width, size_t height, const std::string& caption);
